drivers:stm32:pwm: Change pwm_frequency in _compute_period_ticks() from float to double
Change the datatype of pwm_frequency for improved precision in period tick computation

+++ drivers/platform/stm32/stm32_pwm.c
@@ -56,7 +56,7 @@ static uint32_t _compute_period_ticks(uint32_t (*get_timer_clock)(void),
 				      uint32_t prescaler,
 				      uint32_t period_ns)
 {
-	float pwm_frequency;
+	double pwm_frequency;
 	uint32_t timer_frequency_hz;
 	uint32_t period;
